[Pipeline Refactor][server][OpenAI] Enable OpenAI to use new text gen pipeline (#1477)
* add pipeline.from_config

* update pipeline config; update deepsparses server to use new pipeline

* update tests

* fix imports in server config.py

* fix state update with parse_inputs being added

* update openai server to use new v2 text gen pipeline; disable streaming while unavilable in v2
